// src/core/domain/models/QuestionType.ts
enum QuestionTypes {
  COUNTRY_CAPITAL = 'country-capital',
  COUNTRY_FRONTIER = 'country-frontier',
  COUNTRY_REGION = 'country-region',
  FLAG_COUNTRY = 'flag-country',
  POPULATION_COUNTRY = 'population-country',
  SUBREGION_REGION = 'subregion-region',
}

type QuestionType = QuestionTypes;

export { QuestionType, QuestionTypes };